chore: Stop Extensions from importing FieldDropdown
